Build: decouple `install_build_tools` from `PythonEnvironment`
The method `install_build_tools` is not tied to the Python environment, and it's
more related to the build's director, instead.

I'm moving it to the `BuildDirector` class also because it's needed there when
using `build.commands` as the first step to execute to install the `build.tools`
specified by the user to build their docs.
